backend/requirements.txt
asgiref==3.4.1
astroid==2.9.0
autopep8==1.6.0
bleach==4.1.0
colorama==0.4.4
Django==4.0
django-cors-headers==3.10.1
django-debug-toolbar==3.2.4
django-filter==21.1
django-summernote==0.8.20.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
packaging==21.3
Pillow==8.4.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pylint==2.12.2
pyparsing==3.0.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
webencodings==0.5.1
